Creation controller + ObjectService Fetch Odd
