+++ helpers/leaguesIdentifiers.js
@@ -1,22 +1,22 @@
-module.exports = [
-  { 
+module.exports = {
+  liga: { 
     name: 'soccer_spain_la_liga',
-    alias: 'PD'
+    alias: 'PD',
   },
-  { 
+  serie_a: { 
     name: 'soccer_italy_serie_a',
-    alias: 'SA'
+    alias: 'SA',
   },
-  { 
+  premier_league: { 
     name: 'soccer_epl',
-    alias: 'PL'
+    alias: 'PL',
   },
-  {
+  bundesliga: {
     name: 'soccer_germany_bundesliga',
-    alias: 'BL1'
+    alias: 'BL1',
   },
-  {
+  ligue_1: {
     name: 'soccer_france_ligue_one',
-    alias: 'FL1'
+    alias: 'FL1',
   }
-]
+}
