+++ routes/gamesRoutes.js
@@ -1,311 +1,9 @@
-const axios = require('axios');
-const keys = require('../config/keys');
-const mongoose = require('mongoose');
-
-const Game = mongoose.model('games');
+const GamesControllers = require('../controllers/GamesControllers')
 
 module.exports = (app) => {
 
-  app.get('/api/games/ligue_1', async (req, res) => {
-    const games = await Game.find({sport_key: "soccer_france_ligue_one"})
-    res.send(games)
-  });
-
-  app.get('/api/games/premier_league', async (req, res) => {
-    const games = await Game.find({sport_key: "soccer_epl"})
-    res.send(games)
-  });
-
-  app.get('/api/games/bundesliga', async (req, res) => {
-    const games = await Game.find({sport_key: "soccer_germany_bundesliga"})
-    res.send(games)
-  });
-
-  app.get('/api/games/serie_a', async (req, res) => {
-    const games = await Game.find({sport_key: "soccer_italy_serie_a"})
-    res.send(games)
-  });
-
-  app.get('/api/games/liga', async (req, res) => {
-    const games = await Game.find({sport_key: "soccer_spain_la_liga"})
-    res.send(games)
-  });
+  app.get('/api/games/:league', GamesControllers.fetchDatabase);
 
   // API POUR FETCH LES INFOS, ON NE FAIT LES INFOS QU'UNE FOIS PAR JOUR POUR NE PAS PLOMBER LE QUOTA DE L'API
-  app.get('/api/fetch/games/soccer_france_ligue_one', async (req, res) => { 
-
-    try {
-      const fetchGames = await axios({
-        method: 'get',
-        url: `https://api.the-odds-api.com/v4/sports/soccer_france_ligue_one/odds/?apiKey=${keys.oddsApi}&regions=eu`
-      });
-      
-      const games = fetchGames.data
-      games.forEach(async game => {
-        const { id, home_team, away_team, commence_time, sport_key, bookmakers } = game
-        const unibet = bookmakers.find(bookmaker => bookmaker.key == 'unibet');
-
-        let home_odd = null
-        let away_odd = null
-        let draw_odd = null
-
-        const market = unibet.markets[0]
-
-        market.outcomes.forEach(({name, price}) => {
-          switch(name) {
-            case home_team:
-              return home_odd = price
-            case away_team:
-              return away_odd = price
-            default:
-              return draw_odd = price
-          }
-        })
-
-        const existingGame = await Game.find({_id: id})
-
-        if(existingGame.length > 0) { return null }
-
-        const newGame = new Game({
-          _id: id,
-          home_team: home_team,
-          away_team: away_team,
-          commence_time: commence_time,
-          sport_key: sport_key,
-          home_odd: home_odd,
-          away_odd: away_odd,
-          draw_odd: draw_odd,
-        })
-
-        newGame.save()
-      })
-
-      res.send({ok: 'Les matchs ont été actualisés'})
-
-    } catch(err) {
-      res.status(422).send(err)
-    }
-  });
-
-
-  app.get('/api/fetch/games/soccer_epl', async (req, res) => { 
-
-    try {
-      const fetchGames = await axios({
-        method: 'get',
-        url: `https://api.the-odds-api.com/v4/sports/soccer_epl/odds/?apiKey=${keys.oddsApi}&regions=eu`
-      });
-      
-      const games = fetchGames.data
-      games.forEach(async game => {
-        const { id, home_team, away_team, commence_time, sport_key, bookmakers } = game
-        const unibet = bookmakers.find(bookmaker => bookmaker.key == 'unibet');
-
-        let home_odd = null
-        let away_odd = null
-        let draw_odd = null
-
-        const market = unibet.markets[0]
-
-        market.outcomes.forEach(({name, price}) => {
-          switch(name) {
-            case home_team:
-              return home_odd = price
-            case away_team:
-              return away_odd = price
-            default:
-              return draw_odd = price
-          }
-        })
-
-        const existingGame = await Game.find({_id: id})
-
-        if(existingGame.length > 0) { return null }
-
-        const newGame = new Game({
-          _id: id,
-          home_team: home_team,
-          away_team: away_team,
-          commence_time: commence_time,
-          sport_key: sport_key,
-          home_odd: home_odd,
-          away_odd: away_odd,
-          draw_odd: draw_odd,
-        })
-
-        newGame.save()
-      })
-
-      res.send({ok: 'Les matchs ont été actualisés'})
-
-    } catch(err) {
-      res.status(422).send(err)
-    }
-  });
-
-  app.get('/api/fetch/games/soccer_italy_serie_a', async (req, res) => { 
-
-    try {
-      const fetchGames = await axios({
-        method: 'get',
-        url: `https://api.the-odds-api.com/v4/sports/soccer_italy_serie_a/odds/?apiKey=${keys.oddsApi}&regions=eu`
-      });
-      
-      const games = fetchGames.data
-      games.forEach(async game => {
-        const { id, home_team, away_team, commence_time, sport_key, bookmakers } = game
-        const unibet = bookmakers.find(bookmaker => bookmaker.key == 'unibet');
-
-        let home_odd = null
-        let away_odd = null
-        let draw_odd = null
-
-        const market = unibet.markets[0]
-
-        market.outcomes.forEach(({name, price}) => {
-          switch(name) {
-            case home_team:
-              return home_odd = price
-            case away_team:
-              return away_odd = price
-            default:
-              return draw_odd = price
-          }
-        })
-
-        const existingGame = await Game.find({_id: id})
-
-        if(existingGame.length > 0) { return null }
-
-        const newGame = new Game({
-          _id: id,
-          home_team: home_team,
-          away_team: away_team,
-          commence_time: commence_time,
-          sport_key: sport_key,
-          home_odd: home_odd,
-          away_odd: away_odd,
-          draw_odd: draw_odd,
-        })
-
-        newGame.save()
-      })
-
-      res.send({ok: 'Les matchs ont été actualisés'})
-
-    } catch(err) {
-      res.status(422).send(err)
-    }
-  });
-
-  app.get('/api/fetch/games/soccer_germany_bundesliga', async (req, res) => { 
-
-    try {
-      const fetchGames = await axios({
-        method: 'get',
-        url: `https://api.the-odds-api.com/v4/sports/soccer_germany_bundesliga/odds/?apiKey=${keys.oddsApi}&regions=eu`
-      });
-      
-      const games = fetchGames.data
-      games.forEach(async game => {
-        const { id, home_team, away_team, commence_time, sport_key, bookmakers } = game
-        const unibet = bookmakers.find(bookmaker => bookmaker.key == 'unibet');
-
-        let home_odd = null
-        let away_odd = null
-        let draw_odd = null
-
-        const market = unibet.markets[0]
-
-        market.outcomes.forEach(({name, price}) => {
-          switch(name) {
-            case home_team:
-              return home_odd = price
-            case away_team:
-              return away_odd = price
-            default:
-              return draw_odd = price
-          }
-        })
-
-        const existingGame = await Game.find({_id: id})
-
-        if(existingGame.length > 0) { return null }
-
-        const newGame = new Game({
-          _id: id,
-          home_team: home_team,
-          away_team: away_team,
-          commence_time: commence_time,
-          sport_key: sport_key,
-          home_odd: home_odd,
-          away_odd: away_odd,
-          draw_odd: draw_odd,
-        })
-
-        newGame.save()
-      })
-
-      res.send({ok: 'Les matchs ont été actualisés'})
-
-    } catch(err) {
-      res.status(422).send(err)
-    }
-  });
-
-  app.get('/api/fetch/games/soccer_spain_la_liga', async (req, res) => { 
-
-    try {
-      const fetchGames = await axios({
-        method: 'get',
-        url: `https://api.the-odds-api.com/v4/sports/soccer_spain_la_liga/odds/?apiKey=${keys.oddsApi}&regions=eu`
-      });
-
-      const games = fetchGames.data
-      games.forEach(async game => {
-        const { id, home_team, away_team, commence_time, sport_key, bookmakers } = game
-        const unibet = bookmakers.find(bookmaker => bookmaker.key == 'unibet');
-
-        let home_odd = null
-        let away_odd = null
-        let draw_odd = null
-
-        const market = unibet.markets[0]
-
-        market.outcomes.forEach(({name, price}) => {
-          switch(name) {
-            case home_team:
-              return home_odd = price
-            case away_team:
-              return away_odd = price
-            default:
-              return draw_odd = price
-          }
-        })
-
-        const existingGame = await Game.find({_id: id})
-
-        if(existingGame.length > 0) { return null }
-
-        const newGame = new Game({
-          _id: id,
-          home_team: home_team,
-          away_team: away_team,
-          commence_time: commence_time,
-          sport_key: sport_key,
-          home_odd: home_odd,
-          away_odd: away_odd,
-          draw_odd: draw_odd,
-        })
-
-        newGame.save()
-      })
-
-      res.send({ok: 'Les matchs ont été actualisés'})
-
-    } catch(err) {
-      res.status(422).send(err)
-    }
-  });
-
+  app.get('/api/fetch/games/:league', GamesControllers.fetchApi);
 }
